Rename `Metrics` module to `Metrics::Exported`
Irritatingly, we need to use two different Prometheus libraries as the
one provided through the `govuk_app_config` gem does not support pushed
metrics.

This commit prepares us for this by moving the `Metrics` helper module
to `Metrics::Exported` so we can have a sister module to handle pushed
metrics.
