+++ app/message_processors/publishing_api_message_processor.rb
@@ -1,9 +1,9 @@
 class PublishingApiMessageProcessor
   # Implements the callback interface required by `govuk_message_queue_consumer`
   def process(message)
-    Metrics.increment_counter(:incoming_messages)
+    Metrics::Exported.increment_counter(:incoming_messages)
 
-    Metrics.observe_duration(:total_processing_duration) do
+    Metrics::Exported.observe_duration(:total_processing_duration) do
       document_hash = message.payload.deep_symbolize_keys
       document = PublishingApiDocument.new(document_hash)
       document.synchronize
@@ -11,7 +11,7 @@ class PublishingApiMessageProcessor
 
     message.ack
   rescue StandardError => e
-    Metrics.increment_counter(:message_processing_errors)
+    Metrics::Exported.increment_counter(:message_processing_errors)
 
     # TODO: Consider options for handling errors more granularly, and for differentiating between
     # retriable (e.g. transient connection issue) and fatal (e.g. malformed document on queue)
