+++ app/services/discovery_engine/sync/delete.rb
@@ -11,14 +11,16 @@ module DiscoveryEngine::Sync
       client.delete_document(name: document_name(content_id))
 
       log(Logger::Severity::INFO, "Successfully deleted", content_id:, payload_version:)
-      Metrics.increment_counter(:discovery_engine_requests, type: "delete", status: "success")
+      Metrics::Exported.increment_counter(
+        :discovery_engine_requests, type: "delete", status: "success"
+      )
     rescue Google::Cloud::NotFoundError => e
       log(
         Logger::Severity::INFO,
         "Did not delete document as it doesn't exist remotely (#{e.message}).",
         content_id:, payload_version:,
       )
-      Metrics.increment_counter(
+      Metrics::Exported.increment_counter(
         :discovery_engine_requests, type: "delete", status: "already_not_present"
       )
     rescue Google::Cloud::Error => e
@@ -28,7 +30,9 @@ module DiscoveryEngine::Sync
         content_id:, payload_version:,
       )
       GovukError.notify(e)
-      Metrics.increment_counter(:discovery_engine_requests, type: "delete", status: "error")
+      Metrics::Exported.increment_counter(
+        :discovery_engine_requests, type: "delete", status: "error"
+      )
     end
 
   private
